add MasterSlaveDataSourceRouter
