+++ sharding-core/sharding-core-route/src/main/java/org/apache/shardingsphere/core/route/router/masterslave/MasterSlaveDataSourceRouter.java
@@ -1,0 +1,60 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.shardingsphere.core.route.router.masterslave;
+
+import lombok.RequiredArgsConstructor;
+import org.apache.shardingsphere.api.hint.HintManager;
+import org.apache.shardingsphere.core.rule.MasterSlaveRule;
+import org.apache.shardingsphere.sql.parser.sql.statement.SQLStatement;
+import org.apache.shardingsphere.sql.parser.sql.statement.dml.SelectStatement;
+
+import java.util.ArrayList;
+
+/**
+ * Data source router for master-slave.
+ *
+ * @author zhangliang
+ */
+@RequiredArgsConstructor
+public final class MasterSlaveDataSourceRouter {
+    
+    private final MasterSlaveRule masterSlaveRule;
+    
+    /**
+     * Route.
+     * 
+     * @param sqlStatement SQL statement
+     * @return data source name
+     */
+    public String route(final SQLStatement sqlStatement) {
+        if (isMasterRoute(sqlStatement)) {
+            MasterVisitedManager.setMasterVisited();
+            return masterSlaveRule.getMasterDataSourceName();
+        }
+        return masterSlaveRule.getLoadBalanceAlgorithm().getDataSource(
+                masterSlaveRule.getName(), masterSlaveRule.getMasterDataSourceName(), new ArrayList<>(masterSlaveRule.getSlaveDataSourceNames()));
+    }
+    
+    private boolean isMasterRoute(final SQLStatement sqlStatement) {
+        return containsLockSegment(sqlStatement) || !(sqlStatement instanceof SelectStatement) || MasterVisitedManager.isMasterVisited() || HintManager.isMasterRouteOnly();
+    }
+    
+    private boolean containsLockSegment(final SQLStatement sqlStatement) {
+        return sqlStatement instanceof SelectStatement && ((SelectStatement) sqlStatement).getLock().isPresent();
+    }
+}
