+++ sharding-core/sharding-core-route/src/main/java/org/apache/shardingsphere/core/route/router/masterslave/MasterSlaveRouter.java
@@ -18,18 +18,14 @@
 package org.apache.shardingsphere.core.route.router.masterslave;
 
 import lombok.RequiredArgsConstructor;
-import org.apache.shardingsphere.api.hint.HintManager;
 import org.apache.shardingsphere.core.route.RouteResult;
 import org.apache.shardingsphere.core.route.SQLLogger;
 import org.apache.shardingsphere.core.route.router.DateNodeRouter;
 import org.apache.shardingsphere.core.rule.MasterSlaveRule;
 import org.apache.shardingsphere.sql.parser.SQLParseEngine;
-import org.apache.shardingsphere.sql.parser.sql.statement.SQLStatement;
-import org.apache.shardingsphere.sql.parser.sql.statement.dml.SelectStatement;
 import org.apache.shardingsphere.underlying.route.RouteUnit;
 import org.apache.shardingsphere.underlying.route.SQLUnit;
 
-import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -49,29 +45,12 @@ public final class MasterSlaveRouter implements DateNodeRouter {
     
     @Override
     public RouteResult route(final String sql, final List<Object> parameters, final boolean useCache) {
-        String dataSourceName = getDataSourceName(parseEngine.parse(sql, useCache));
+        String dataSourceName = new MasterSlaveDataSourceRouter(masterSlaveRule).route(parseEngine.parse(sql, useCache));
         if (showSQL) {
             SQLLogger.logSQL(sql, dataSourceName);
         }
         RouteResult result = new RouteResult(null);
         result.getRouteUnits().add(new RouteUnit(dataSourceName, new SQLUnit(sql, parameters)));
         return result;
-    }
-    
-    private String getDataSourceName(final SQLStatement sqlStatement) {
-        if (isMasterRoute(sqlStatement)) {
-            MasterVisitedManager.setMasterVisited();
-            return masterSlaveRule.getMasterDataSourceName();
-        }
-        return masterSlaveRule.getLoadBalanceAlgorithm().getDataSource(
-                masterSlaveRule.getName(), masterSlaveRule.getMasterDataSourceName(), new ArrayList<>(masterSlaveRule.getSlaveDataSourceNames()));
-    }
-    
-    private boolean isMasterRoute(final SQLStatement sqlStatement) {
-        return containsLockSegment(sqlStatement) || !(sqlStatement instanceof SelectStatement) || MasterVisitedManager.isMasterVisited() || HintManager.isMasterRouteOnly();
-    }
-
-    private boolean containsLockSegment(final SQLStatement sqlStatement) {
-        return sqlStatement instanceof SelectStatement && ((SelectStatement) sqlStatement).getLock().isPresent();
     }
 }
